+++ dotnet/corecomponents/net.atos.daf.ct2.alert/entity/NotificationLimit.cs
@@ -13,7 +13,7 @@ namespace net.atos.daf.ct2.alert.entity
         public string NotificationPeriodType { get; set; }
         public int PeriodLimit { get; set; }
         public string State { get; set; }
-        public int CreatedAt { get; set; }
-        public int ModifiedAt { get; set; }
+        public long CreatedAt { get; set; }
+        public long ModifiedAt { get; set; }
     }
 }
